Fix initiatives transl. storage in BRS integration
Some translations were stored with the wrong original property names
(`q2` and `q3`), but we agreed to use new more meaningful names for
them, instead: `title` and `summary`, respectively. Therefore, since
the BRS integration was expecting the names to match, we have needed to
add some exceptions to transform the data when storing it for
translations.