Impl remove-user-from-channel RocketChat API method
[Re #1590]

* We are using the *.kick. For private channels RocketChat applies the
same logic as with retrieving members for a private channel, the Admin
callee needs to be part of that private group. This is a well-known
issue in the RocketChat code base but still not solved. For the time
being we have to make sure the Admin user we use to call the
RocketChat API, joins every new channel created.