Support for private tags in DB model
[Re #1619]

Added new `private` boolean property with default `false` value for the
Tag entity, to support what is described in the issue. This new
property will not be enforced but currently existing tags will use
`false` value by default and also whenever this property is not
explictly set, as that will be the usual case
(publicly-available tags).