Add create tag POST endpoint for super admins
[Re #1619]

Added new endpoint for super admins to create a tag within a given tag
category.

This endpoint provides some default values for `review_status` property,
as it is meant to be used by super admins where the tags need to be
approved by default.

Besides, the new endpoint also gives meaningful errors when trying to
refer to a non-existing tag category. For this purpose, we have needed
to refactor the common tags creation code to throw specific exceptions
when needed.

On the other hand, we have refactored the shared create tags code to
allow to provide `private` property for each created tag, using `false`
default value when it is not provided.

For this endpoint we are using the existing Tag entity schema,
overriding some settings to adjust the schema to the endpoint. The idea
is to re-use this schema for the PUT endpoint and such.