Merge Res. Detail and Initiative GET endpoints
[Re #1615]

As we have agreed, we are getting rid of the redundant `/api/initiative/
{id}` GET endpoint to provide all the information in a single endpoint:
the current one already used for resource detail GET operation.

In this case, we have needed to change the shared topic generation SQL
code to get all the properties from initiative entity so then we can
merge it with the other renamed and transformed properties without the
need to fetch the data from the DB twice.

During this process we have got rid of legacy and not-existing column
aliases used in the shared query to get initiatives data for Browse and
Detail API.