Add notif. when accepting priv. channel join reqs
[Re #1625]

Now we notify (via email) the user when it is accepted after he has
requested to join a Forum's private channel.

In order to simplify the process, we expect the `channel_name` to be
provided as part of the endpoint used for the point above, since this
information is already available in the front-end, so we avoid to get
it in an additional request to Rocket.chat.