Fix Organisation tags association
[Re #1629]

We were misusing the result of the DB query to update an organisation,
as we were using the affected rows number as it was the organisation's
updated id, that was being passed to the function to associate tags
with the organisation. Hence, we have ended-up associating those tags
to the organisation with id = `1` (UNEP).

Besides, since we no longer have organisations with id `-1` to handle
GPML membership process, we don't need at all to update an id of an
organisation there, which is quite risky as well. So, now we avoid that
and we also make sure we use the entity's domain properties, removing
the ones that are not expected to be updated in each case.

Finally, we also provide the right `organisation` id coming from the
Path params of the endpoint for making the tags associations.

In order to perform the mentioned changes easier, we have refactored the
update code to be dynamic and not harcode the updated params. For that
we have followed a similar approach as for the resource update query,
adding DB functions declarations as well.