Fix Organisation logo update
[Re #1629]

We were not handling well logo updates, as we were expecting to always
get it, so if we did not want to update it as part of the rest of the
organisation data, the backend would delete any previous logo image.
Now we make sure we try to delete or replace the logo if we explictly
get this property as part of the update data from the different
endpoints involved.

On the other hand, we have introduced a bug in the previous commit,
since we were expecting a wrong format for the logo coming in the API
requests to update organisation data. We were expecting a URL but that
does not make any sense, since it should be a base64 string, as this is
the content of the logo that is going to be stored. The domain entity's
schema was wrong when it was built and we did not notice that when we
started using it here.