Provide `private` property of tags in GET Resource Detail endpoint
[Re #1619]

We have been requested to load the mentioned property among the tags
data related to the target resource data obtained in the GET Resource
Details-related endpoint.

As a side effect, we will also load it in other use-cases, but that is
not harmful and it can be even better for the future.