Add invitation type column to support PS user invitation
[Re #1611]

* The invitiation type is to distinguish which type of invitation
it's. As of now we have one type of invitation in production which is
"expert" and with PS we need another type of invitation. As experts
have extra logic when the user actually accepts the invitation, the
"INVITED" status is not enough. So this "type" column will give us the
necessary information to distinguish both cases and more in the future.