Fix expertise tags in `/api/organisation/{id}` GET endpoint
[Re #1636]

As explained in the issue, we need to return all the tags related to the
target organisation, for the endpoint mentioned above. Hence, we have
removed the legacy DB query to use the generic one for the resource
details get functionality. Obviously we no longer get tags grouped by
category and we avoid the strange random get of the first group of
tags, so we just return a collection with the ids of all the found
tags.

In order to re-use the shared mentioned get resource tag's DB query, we
have needed to modify it so we can optionally pass a filter to be used
in the current places it was being used, so only approved tags were
considered there, while we can use the query without filters for the
use-case we are dealing with here.