Fix emptying tags in resource/org edition
[Re #1638]

As described in the issue, now we check in the affected API endpoints
whether we get the `tags` key at all in the payload, so we know if we
need to remove all the tag relationships, add/remove some of them or we
just ignore tags during the update (in case we do not receive such
property at all).

We have needed to update the generic tag update code to allow this
use-case.