Add new endp. for GET Organisations in Stak. map
[Re #1611]

As we needed to get Plastic Strategy Workflow feature-specific new
information related to organisations, which includes advance sorting
and filtering capabilities, we have decided to implement a new endpoint
just for that, which will be also useful in the future for other
use-cases.

The reason not to use Browse API for this was to avoid adding more
complexity to that query, as in the past we removed organisations as
resource type to be provided through Browse API, in order to keep
performance and complexity in the right balance. On the other hand,
community members related endpoint was mixing stakeholders with the
organisations data and it would be a mess as well to re-use this
endpoint for the Stakeholder Map view in PSW.

PSW Bookmarking related capabilities are yet to be added to this new
endpoint.