Support PS bookmarks in GET Organisations endpoint
[Re #1611]

Now we support Plastic Strategy Workflow's bookmarking feature in this
endpoint.

Regarding the PS bookmarking information, now we can add that data to
each organisation queried from the list, by using
`ps_country_iso_code_a2` query param.

Once we get PS Bookmark-related information, we can further filter those
bookmarks by providing `ps_bookmark_sections_keys` property, as a list
of PS sub-section keys where the bookmarks are related to.

On the other hand, we can order the obtained organisations by providing
`ps_bookmarked` value in the `order_by` param value, so we can put
first the bookmarked organisations among the results when `true` value
is provided. In order to do the opposite, we need to provide
`not_ps_bookmarked` value for the `order_by` property, as we are
ignoring the `descending` query param, since it would be confusing in
this case.

Finally, we can filter the organisations by providing a boolean value
for the `ps_bookmarked` query param, in order to keep bookmarked or
non-bookmarked organisations in the list. This filter relies on the
previous mentioned `ps_bookmark_sections_keys` properties and
`ps_country_iso_code_a2` to be provided as well.