Add support for program. Badge creation
[Re #1611]

Added new programmatic API endpoint for creating a Badge. As always,
this endpoint is restricted for internal usage.

We are not using JDBC transactions in the process since we need to
perform side-effectful operations that cannot be just undone by
aborting JDBC transactions, so we are explictly doing rollbacks when
some steps cannot be performed and some others have been already
committed.