Implement GET Badge endpoint
[Re #1611]

The new endpoint expects a `name` path parameter to find the target
Badge by its unique name, as this is the most useful way to use this
endpoint. As part of the data fetch the endpoint provides the content
file's URL, retrieved from GCS. The `content-file-id` parameter is not
used after getting the file url, so we remove it from the result.