Implement POST Badge Assign/Unassign endpoint
[Re #1611]

This new endpoint behaves in a similar way as the PS Bookmark-related
endpoint, in the sense that it can be used for both assigning or
unassigning a badge from a given entity type, providing a flag to know
the action and other details to identify the resource type, id and the
badge id as well.

In this case we are handling the entity type in an easier and safer way
than it is done with PS Bookmarking, as we are avoiding relying on `_`
separator to infer the entity type. Hence, this bug will need to be
fixed in PS Bookmark handling, but that is a separated issue.

As a side note, we are sharing a single param for both GET and POST
Assign Badge endpoints, in order to avoid route conflicts, so the path
param can be an integer id or a name string in each case, what could be
confusing, as one can imagine that you can provide name or id for both
endpoints, but it doesn't work like that (as explained in the Swagger
params.