Support Badges data in Browse API GET endpoint
[Re #1611]

Now we load assigned badges metadata to each resource fetched with this
endpoint. However, we are just providing relationship-related data and
Badge's related extra data, such the File content URL, needs to be
fetched using the individual `/api/badge` GET endpoint.

For now we do not offer filtering or sorting capabilities regarding
badges. That can also comes in the future.