Support getting Badge by id
[Re #1611]

We load badge id as part of the relationship data of badge-resource
assignments, so it make sense we can use the badge id to obtain all of
its information, instead of just allowing to get it by its unique
name.

Added more clarifications about the shared context `id-or-name`
parameter, as for assignments we only expect the badge id to be
provided, as that is easier than using the name, for that case.