Support using Badge name in /api/badge/assing endp.
[Re #1611]

We need to support assinging and unassigning badge by its unique name
since it is more convinient for FE to avoid relying on extra GET call
for getting the badge ids.

We have decided to support both name and id properties here because it's
easier and we can share same specs for the GET and POST Assign Badge
endpoints, regarding path params.