Create chat account when adding PS team member if required
[Re #1611]

* Every user added to a plastic strategy needs to have a chat account
enabled in order to add them to the PS channel. Therefore, I added an
extra step to create the user account in RocketChat if it doesn't have
one already. Since creating the "Forum" (i.e. RocketChat) account is
completely transparent for the user using the platform, this is also
something that only concerns the BE.