Support strengths prop filtering by PS code in Organisations GET endpoint
[Re #1611]

We need to filter the strengths count according to the provided PS iso
code from the params, so only those bookmarked initiatives for the
given PS are taking into account when the PS iso code is provided. In
this case, we do not count again same bookmarks for different sections,
as it is only relevant the PS used to bookmark those initiatives.