Fix community API text search methodology
* We were using the full text search functionality for PostgreSQL
which is not very helpful for our use case where we might have content
in different languages or similar words (such similar nouns) that
wouldn't be transformed in equal lexemes for doing text search. A more
reliable approach for this situation is to use ILIKE which will match
the exact query word at any point of the search query.