Allow total tags removal from stakeholders
[Re #1734]

As explained in the issue, the generic user (stakeholder) tag
assignation machinery was not ready to allow to remove all the tags of
a given relation-category (offering, seeking or expertise), so now we
have refactored the code to allow so. Since, the tags assignation is a
partial thing (per category), the fix is quite more ellaborated, as we
need to keep the old tags that are not part of the target category, so
those are not overriden.