Fix logic for transnational_counts_per_country CTE in map-counts
* The idea here is to get those resources that have
"geo_coverage_type" = "transnational", take their transnational
values ("geo_coverage_country_groups") and retrieve the countries
associated to each transnational. This way we link a transnational
resource to all countries that are part of a specific transnational
group. For example, if we have an Event resource which its
"geo_coverage_type" is "transnational" and the value is "Europe", we
should add +1 to the counter of all countries that are part of
"Europe" country group.
* Before we were matching all possible values of "geo_coverage" by
using the "geo_coverage_values" which has a mix of
"geo_coverage_countries" and "geo_coverage_country_groups" which is
not right. We just want to get the country groups of the resource and
link it to the countries that are part of those country groups.